new try

+++ hashmap.c
@@ -127,7 +127,7 @@ Pair * firstMap(HashMap * map)
 Pair * nextMap(HashMap * map) 
 {
   map->current = (map->current+1)%map->capacity;
-  long posicion = map->current;
+  int posicion = map->current;
   while(posicion < map->capacity && (map->buckets[posicion]==NULL || map->buckets[posicion]->key==NULL))
       {
         posicion = (posicion+1)%map->capacity;
